+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/ir/irFunctionAccessExpression.kt
@@ -2,13 +2,11 @@
 
 package com.javiersc.kotlin.compiler.extensions.ir
 
-import com.javiersc.kotlin.compiler.extensions.common.fqName
 import org.jetbrains.kotlin.ir.declarations.IrParameterKind
 import org.jetbrains.kotlin.ir.declarations.IrValueParameter
 import org.jetbrains.kotlin.ir.expressions.IrConstructorCall
 import org.jetbrains.kotlin.ir.expressions.IrExpression
 import org.jetbrains.kotlin.ir.expressions.IrFunctionAccessExpression
-import org.jetbrains.kotlin.ir.util.hasAnnotation
 import org.jetbrains.kotlin.ir.util.kotlinFqName
 import org.jetbrains.kotlin.name.CallableId
 import org.jetbrains.kotlin.name.FqName
@@ -25,12 +23,6 @@ public inline val IrFunctionAccessExpression.name: Name
 
 public inline val IrFunctionAccessExpression.annotations: List<IrConstructorCall>
     get() = symbol.owner.annotations
-
-public inline fun <reified T : Annotation> IrFunctionAccessExpression.hasAnnotation(): Boolean =
-    annotations.hasAnnotation(fqName<T>())
-
-public inline fun IrFunctionAccessExpression.hasAnnotation(annotation: FqName): Boolean =
-    annotations.hasAnnotation(annotation)
 
 public val IrFunctionAccessExpression.extensionReceiver: IrExpression?
     get() = getExtensionReceiverOrNull()